requirements.txt
fastapi
uvicorn
websockets
python-multipart
passlib[bcrypt]
python-jose
bcrypt==3.2.2
sqlalchemy
python-dotenv
